Goal 1 completed
The LIST command will now print a list of all possible <From> <To> pairs
for the pieces of the current player.

Wrote a abstract method 'getValidMoves' in the piece class. Each piece
has an implementation of this method which returns a set of possible and
valid moves for that piece.